+++ numba/parfors/parfor.py
@@ -29,7 +29,7 @@ from numba.core.imputils import impl_ret_untracked
 import numba.core.ir
 from numba.core import types, typing, utils, errors, ir, analysis, postproc, rewrites, typeinfer, config, ir_utils
 from numba import prange, pndindex
-from numba.np import npdatetime
+from numba.np.npdatetime_helpers import datetime_minimum, datetime_maximum
 from numba.np.numpy_support import as_dtype, numpy_version
 from numba.core.typing.templates import infer_global, AbstractTemplate
 from numba.stencils.stencilparfor import StencilPass
@@ -124,25 +124,6 @@ class internal_prange(object):
     def __new__(cls, *args):
         return range(*args)
 
-def _datetime_minimum(a, b):
-    pass
-
-def _datetime_maximum(a, b):
-    pass
-
-@infer_global(_datetime_minimum)
-@infer_global(_datetime_maximum)
-class _DatetimeMinMaxInfer(AbstractTemplate):
-    def generic(self, args, kws):
-        assert not kws
-        assert len(args) == 2
-        assert isinstance(args[0], (types.NPDatetime, types.NPTimedelta))
-        assert isinstance(args[1], (types.NPDatetime, types.NPTimedelta))
-        return signature(args[0], *args)
-
-lower_builtin(_datetime_minimum, types.VarArg(types.Any))(npdatetime.datetime_minimum_impl)
-
-lower_builtin(_datetime_maximum, types.VarArg(types.Any))(npdatetime.datetime_maximum_impl)
 
 def min_parallel_impl(return_type, arg):
     # XXX: use prange for 1D arrays since pndindex returns a 1-tuple instead of
@@ -158,7 +139,7 @@ def min_parallel_impl(return_type, arg):
                 min_checker(len(in_arr))
                 val = numba.cpython.builtins.get_type_max_value(in_arr.dtype)
                 for i in numba.parfors.parfor.internal_prange(len(in_arr)):
-                    val = _datetime_minimum(val, in_arr[i])
+                    val = datetime_minimum(val, in_arr[i])
                 return val
         else:
             def min_1(in_arr):
@@ -190,7 +171,7 @@ def max_parallel_impl(return_type, arg):
                 max_checker(len(in_arr))
                 val = numba.cpython.builtins.get_type_min_value(in_arr.dtype)
                 for i in numba.parfors.parfor.internal_prange(len(in_arr)):
-                    val = _datetime_maximum(val, in_arr[i])
+                    val = datetime_maximum(val, in_arr[i])
                 return val
         else:
             def max_1(in_arr):
@@ -3654,8 +3635,8 @@ def supported_reduction(x, func_ir):
         if callname in [
             ('max', 'builtins'),
             ('min', 'builtins'),
-            ('_datetime_minimum', 'numba.parfors.parfor'),
-            ('_datetime_maximum', 'numba.parfors.parfor'),
+            ('datetime_minimum', 'numba.np.npdatetime_helpers'),
+            ('datetime_maximum', 'numba.np.npdatetime_helpers'),
         ]:
             return True
     return False
